requirements.txt
flask
pandas
numpy
faiss-cpu
sentence-transformers
lightgbm
scikit-learn